Use size_t in CWallet::MarkDestinationsDirty
Summary: As per title.

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7835

+++ src/wallet/wallet.cpp
@@ -3565,7 +3565,8 @@ void CWallet::MarkDestinationsDirty(
         if (wtx.m_is_cache_empty) {
             continue;
         }
-        for (unsigned int i = 0; i < wtx.tx->vout.size(); i++) {
+
+        for (size_t i = 0; i < wtx.tx->vout.size(); i++) {
             CTxDestination dst;
 
             if (ExtractDestination(wtx.tx->vout[i].scriptPubKey, dst) &&
